LayoutTests/storage/indexeddb/keypath-basics-expected.txt

Test for valid and invalid keypaths

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
store = db.createObjectStore('name')
PASS store.keyPath is null
Deleted all object stores.
store = db.createObjectStore('name', {keyPath: null})
PASS store.keyPath is null
index = store.createIndex('name', null)
PASS index.keyPath is 'null'
Deleted all object stores.
store = db.createObjectStore('name', {keyPath: undefined})
PASS store.keyPath is null
index = store.createIndex('name', undefined)
PASS index.keyPath is 'undefined'
Deleted all object stores.
store = db.createObjectStore('name', {keyPath: ''})
PASS store.keyPath is ''
index = store.createIndex('name', '')
PASS index.keyPath is ''
Deleted all object stores.
store = db.createObjectStore('name', {keyPath: 'foo'})
PASS store.keyPath is 'foo'
index = store.createIndex('name', 'foo')
PASS index.keyPath is 'foo'
Deleted all object stores.
store = db.createObjectStore('name', {keyPath: 'foo.bar.baz'})
PASS store.keyPath is 'foo.bar.baz'
index = store.createIndex('name', 'foo.bar.baz')
PASS index.keyPath is 'foo.bar.baz'
Deleted all object stores.
Deleted all object stores.
globalKeyPath = '[]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '["foo"]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '["foo", "bar"]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '["", ""]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '[1.0, 2.0]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '[["foo"]]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '["foo", ["bar"]]'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = ' '
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo '
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo bar'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo. bar'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo .bar'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo..bar'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '+foo'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo%'
Expecting exception from db.createObjectStore('name', {keyPath: globalKeyPath})
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = ' '
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo '
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo bar'
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo. bar'
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo .bar'
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo..bar'
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = '+foo'
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
globalKeyPath = 'foo%'
store = db.createObjectStore('storeName')
Expecting exception from store.createIndex('name', globalKeyPath)
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
Deleted all object stores.
PASS successfullyParsed is true

TEST COMPLETE
